+++ pgcom/exc.py
@@ -1,0 +1,12 @@
+class ExecutionError(Exception):
+    """Raised when query execution fails.
+    """
+
+    pass
+
+
+class CopyError(Exception):
+    """Raised when COPY FROM fails.
+    """
+
+    pass